[mobile] 自分に送信ボタンタップ時の動作を実装
